removed example Handler from core package moved serverless config to sample package
