+++ ktor-server-lambda-sample/serverless.yml
@@ -6,11 +6,11 @@ provider:
   region: eu-central-1
 
 package:
-  artifact: target/example-0.0.1-aws.jar
+  artifact: target/ktor-server-lambda-sample-0.0.1-aws.jar
 
 functions:
   hello:
-    handler: com.ktor.HandlerKt::handler
+    handler: com.mercateo.oss.HandlerKt::handler
     timeout: 30
     events:
       - http:
